Attempting damage

+++ huntduck/Assets/Scripts/PlayerWeapon.cs
@@ -5,10 +5,10 @@ using UnityEngine;
 
 // Unity will know how to save & load this class so we can update the range and damage in the inspector
 [System.Serializable]
-public class PlayerWeapon 
+public class PlayerWeapon
 {
     public string name = "Gun";
 
-    public float damage = 10f;
+    public int damage = 10;
     public float range = 100f;
 }
